Move CustomTabsHelper to app module
