+++ app/src/main/java/io/github/droidkaigi/confsched2018/util/CustomTabsHelper.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.chromium.customtabsclient.shared;
+package io.github.droidkaigi.confsched2018.util;
 
 import android.content.Context;
 import android.content.Intent;
