+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':model', ':customtabs'
+include ':app', ':model'
